refactor(DatasetExplorer): extract chart components into separate files (#120)
Extract 8 render functions (~2,000 lines) from DatasetExplorer.tsx into
11 standalone React components using ChartContext for shared state.

- Add ChartContext with ~45 shared dependencies (aggregation, filters,
  count-by, colors, range filters, view preferences, dashboard, survival)
- Extract ChartHeader, CountIndicator, FilterMenu, NumericFilterMenu
- Extract PieChart, BarChart, TableViewChart, HistogramChart,
  SurvivalChart, MapChart
- Add component tests with mock context provider (testUtils)
- DatasetExplorer.tsx reduced from 5,489 to 3,513 lines (36% reduction)
- All 141 tests pass, build succeeds cleanly
