+++ client/src/pages/DatasetExplorer/DatasetExplorer.tsx
@@ -1,7 +1,5 @@
 import React, { useState, useEffect, useRef, useMemo } from 'react'
 import { useParams, useNavigate, useLocation } from 'react-router-dom'
-import Plot from 'react-plotly.js'
-import type { PlotMouseEvent, PlotSelectionEvent } from 'plotly.js'
 import SafeHtml from '../../components/SafeHtml'
 import api from '../../services/api'
 import type { MetricPathSegment } from '../../types'
@@ -21,14 +19,19 @@ import { encodeState, decodeState } from '../../utils/urlHelpers'
 import {
   type CountBySelection
 } from '../../utils/presetHelpers'
-import { getStateCode, normalizeStateName } from '../../data/us-states'
 import { useFilterPresets } from './hooks/useFilterPresets'
 import { useViewPreferences } from './hooks/useViewPreferences'
 import { useDashboard } from './hooks/useDashboard'
+import { ChartProvider } from './components/ChartContext'
+import { PieChart } from './components/PieChart'
+import { BarChart } from './components/BarChart'
+import { TableViewChart } from './components/TableViewChart'
+import { HistogramChart } from './components/HistogramChart'
+import { SurvivalChart } from './components/SurvivalChart'
+import { MapChart } from './components/MapChart'
 import {
   MAX_PIE_CATEGORIES,
   CHART_LABEL_STORAGE_PREFIX,
-  TABLE_SCOPE_KEY,
   DASHBOARD_SCOPE_KEY,
   CACHE_TTL_MS,
   CACHE_MAX_ENTRIES_PER_TABLE,
@@ -36,7 +39,6 @@ import {
   persistChartOverrides,
   loadChartOverrides,
   type ColumnMetadata,
-  type CategoryCount,
   type NumericStats,
   type HistogramBin,
   type SurvivalCurvePoint,
@@ -1086,24 +1088,6 @@ function DatasetExplorer() {
     )
   }
 
-  const renderTableCountIndicator = (tableName: string, columnName: string, cacheKey: string) => {
-    const indicatorColor = getCountIndicatorColor(tableName, cacheKey)
-    const borderColor = getCountByTableColor(tableName, cacheKey)
-    const label = getCountByLabelFromCacheKey(tableName, cacheKey)
-    const options = getCountByOptions(tableName)
-    const currentValue = cacheKey
-    return renderCountIndicator({
-      menuKey: `${TABLE_SCOPE_KEY}:${tableName}.${columnName}`,
-      indicatorColor,
-      borderColor,
-      label,
-      options,
-      currentValue,
-      buttonLabel: `Change count-by for ${tableName}.${columnName}`,
-      onSelect: value => handleChartCountOverrideChange(tableName, columnName, value)
-    })
-  }
-
   const renderDashboardCountIndicator = (
     chartIndex: number,
     tableName: string,
@@ -1354,151 +1338,6 @@ function DatasetExplorer() {
     })
   }
 
-  const renderFilterMenu = (
-    tableName: string,
-    columnName: string,
-    categories?: CategoryCount[],
-    cacheKeyOverride?: string
-  ) => {
-    const cacheKey = cacheKeyOverride ?? getEffectiveCacheKeyForChart(tableName, columnName)
-    const menuOpen =
-      activeFilterMenu?.tableName === tableName &&
-      activeFilterMenu.columnName === columnName &&
-      activeFilterMenu.countKey === cacheKey
-    if (!menuOpen) return null
-
-    const aggregation = getAggregation(tableName, columnName, cacheKey)
-    if (!aggregation || !categories || categories.length === 0) return null
-
-    const metricLabels = getMetricLabels(aggregation)
-
-    const columnHasFilter = hasColumnFilter(columnName, cacheKey)
-
-    // Check if the current filter for this column has NOT wrapper
-    const currentFilter = filters.find(f => {
-      const actualF = unwrapNot(f)
-      if (!actualF || getFilterColumn(actualF) !== columnName) return false
-      return getFilterCountKey(f) === cacheKey
-    })
-    const isNot = !!currentFilter?.not
-
-    // Toggle NOT for this column's filter
-    const toggleColumnNot = () => {
-      setFilters(prev => {
-        const idx = prev.findIndex(f => {
-          const actualF = unwrapNot(f)
-          if (!actualF || getFilterColumn(actualF) !== columnName) return false
-          return getFilterCountKey(f) === cacheKey
-        })
-        if (idx === -1) return prev
-
-        const updated = [...prev]
-        const filter = prev[idx]
-        if (filter.not) {
-          // Remove NOT wrapper
-          updated[idx] = filter.not
-        } else {
-          // Add NOT wrapper
-          updated[idx] = { not: filter }
-        }
-        return updated
-      })
-    }
-
-    // Parent counting can double-book entities across wedges, so prefer bars
-    if (aggregation.metric_type === 'parent') {
-      return null
-    }
-
-    return (
-      <div
-        style={{
-          position: 'absolute',
-          top: '28px',
-          right: 0,
-          zIndex: 10,
-          background: 'white',
-          border: '1px solid #ddd',
-          borderRadius: '6px',
-          boxShadow: '0 4px 12px rgba(0,0,0,0.12)',
-          padding: '0.5rem',
-          maxHeight: '200px',
-          overflowY: 'auto',
-          minWidth: '140px',
-          display: 'flex',
-          flexDirection: 'column',
-          gap: '0.25rem'
-        }}
-        onClick={(event) => event.stopPropagation()}
-      >
-        <div style={{ display: 'flex', gap: '0.25rem' }}>
-          <button
-            onClick={() => clearColumnFilter(tableName, columnName, cacheKey)}
-            style={{
-              border: 'none',
-              background: columnHasFilter ? '#1976D2' : '#eee',
-              color: columnHasFilter ? 'white' : '#555',
-              borderRadius: '4px',
-              padding: '0.25rem 0.5rem',
-              fontSize: '0.7rem',
-              cursor: columnHasFilter ? 'pointer' : 'default',
-              opacity: columnHasFilter ? 1 : 0.6,
-              flex: 1
-            }}
-            disabled={!columnHasFilter}
-          >
-            Reset
-          </button>
-          <button
-            onClick={toggleColumnNot}
-            style={{
-              border: 'none',
-              background: isNot ? '#333' : '#f0f0f0',
-              color: isNot ? 'white' : '#555',
-              borderRadius: '4px',
-              padding: '0.25rem 0.5rem',
-              fontSize: '0.7rem',
-              cursor: columnHasFilter ? 'pointer' : 'default',
-              opacity: columnHasFilter ? 1 : 0.6,
-              fontWeight: isNot ? 'bold' : 'normal'
-            }}
-            disabled={!columnHasFilter}
-            title={isNot ? 'Remove NOT' : 'Add NOT'}
-          >
-            ¬
-          </button>
-        </div>
-        <div style={{ borderBottom: '1px solid #eee', margin: '0.25rem 0' }} />
-        {categories.map(category => {
-          const rawValue = normalizeFilterValue(category.value)
-          const label = category.display_value ?? (category.value === '' ? '(Empty)' : String(category.value))
-          const active = isValueFiltered(columnName, rawValue, cacheKey)
-
-          return (
-            <button
-              key={`${tableName}-${columnName}-${label}`}
-              onMouseDown={event => event.preventDefault()}
-              onClick={() => toggleFilter(columnName, rawValue, tableName, cacheKey)}
-              style={{
-                border: active ? '1px solid #1976D2' : '1px solid #ccc',
-                background: active ? '#E3F2FD' : '#fafafa',
-                color: active ? '#0D47A1' : '#444',
-                borderRadius: '999px',
-                padding: '0.25rem 0.5rem',
-                fontSize: '0.7rem',
-                cursor: 'pointer',
-                textAlign: 'left'
-              }}
-              title={`${label} (${category.count} ${metricLabels.short})`}
-            >
-              {label}
-            </button>
-          )
-        })}
-      </div>
-    )
-  }
-
   useEffect(() => {
     if (!activeFilterMenu) return
     const { tableName, columnName, countKey } = activeFilterMenu
@@ -1711,271 +1550,6 @@ function DatasetExplorer() {
     return filtered.length > 0 ? filtered : histogram
   }
 
-  const renderNumericFilterMenu = (
-    tableName: string,
-    columnName: string,
-    histogram?: HistogramBin[],
-    stats?: NumericStats,
-    cacheKeyOverride?: string
-  ) => {
-    const cacheKey = cacheKeyOverride ?? getEffectiveCacheKeyForChart(tableName, columnName)
-    const menuOpen =
-      activeFilterMenu?.tableName === tableName &&
-      activeFilterMenu.columnName === columnName &&
-      activeFilterMenu.countKey === cacheKey
-    if (!menuOpen) return null
-
-    const bins = histogram ?? []
-    const aggregation = getAggregation(tableName, columnName, cacheKey)
-    if (!aggregation) return null
-
-    const metricLabels = getMetricLabels(aggregation)
-    const key = rangeKey(tableName, columnName, cacheKey)
-    const range = customRangeInputs[key] || { min: stats && stats.min !== null ? String(stats.min) : '', max: stats && stats.max !== null ? String(stats.max) : '' }
-    const columnHasFilter = hasColumnFilter(columnName, cacheKey)
-    const selectedRanges = rangeSelections[key] ?? []
-    const customRanges = selectedRanges.filter(range => !bins.some(bin => rangesEqual(range, { start: bin.bin_start, end: bin.bin_end })))
-    const minDisplay = stats && stats.min !== null ? formatRangeValue(stats.min) : '–'
-    const maxDisplay = stats && stats.max !== null ? formatRangeValue(stats.max) : '–'
-    const medianDisplay = stats && stats.median !== null ? formatRangeValue(stats.median) : '–'
-    const stdDisplay = stats && stats.stddev !== undefined && stats.stddev !== null ? stats.stddev.toFixed(2) : '–'
-
-    const minValue = Number(range.min)
-    const maxValue = Number(range.max)
-    const hasValidRange =
-      range.min.trim() !== '' &&
-      range.max.trim() !== '' &&
-      Number.isFinite(minValue) &&
-      Number.isFinite(maxValue) &&
-      minValue <= maxValue
-
-    // Check if the current filter for this column has NOT wrapper
-    const currentFilter = filters.find(f => {
-      const actualF = unwrapNot(f)
-      if (!actualF || getFilterColumn(actualF) !== columnName) return false
-      return getFilterCountKey(f) === cacheKey
-    })
-    const isNot = !!currentFilter?.not
-
-    // Toggle NOT for this column's filter
-    const toggleColumnNot = () => {
-      setFilters(prev => {
-        const idx = prev.findIndex(f => {
-          const actualF = unwrapNot(f)
-          if (!actualF || getFilterColumn(actualF) !== columnName) return false
-          return getFilterCountKey(f) === cacheKey
-        })
-        if (idx === -1) return prev
-
-        const updated = [...prev]
-        const filter = prev[idx]
-        if (filter.not) {
-          // Remove NOT wrapper
-          updated[idx] = filter.not
-        } else {
-          // Add NOT wrapper
-          updated[idx] = { not: filter }
-        }
-        return updated
-      })
-    }
-
-    return (
-      <div
-        style={{
-          position: 'absolute',
-          top: '28px',
-          right: 0,
-          zIndex: 10,
-          background: 'white',
-          border: '1px solid #ddd',
-          borderRadius: '6px',
-          boxShadow: '0 4px 12px rgba(0,0,0,0.12)',
-          padding: '0.5rem',
-          maxHeight: '260px',
-          overflowY: 'auto',
-          minWidth: '180px',
-          display: 'flex',
-          flexDirection: 'column',
-          gap: '0.35rem'
-        }}
-        onClick={(event) => event.stopPropagation()}
-      >
-        {stats && (
-          <>
-            <div style={{ fontSize: '0.7rem', color: '#555', display: 'flex', justifyContent: 'space-between', gap: '0.5rem' }}>
-              <span>Min: {minDisplay}</span>
-              <span>Max: {maxDisplay}</span>
-            </div>
-            <div style={{ fontSize: '0.7rem', color: '#555', display: 'flex', justifyContent: 'space-between', gap: '0.5rem' }}>
-              <span>Median: {medianDisplay}</span>
-              <span>Std: {stdDisplay}</span>
-            </div>
-          </>
-        )}
-        {(() => {
-          const aggregation = getAggregation(tableName, columnName, cacheKey)
-          const metricLabels = getMetricLabels(aggregation)
-          const nullCount = aggregation?.null_count ?? 0
-          if (nullCount === 0) return null
-
-          const nullActive = isValueFiltered(columnName, '', cacheKey)
-          return (
-            <>
-              <div style={{ borderBottom: '1px solid #eee', margin: '0.25rem 0' }} />
-              <button
-                onMouseDown={event => event.preventDefault()}
-                onClick={() => toggleFilter(columnName, '', tableName, cacheKey)}
-                style={{
-                  border: nullActive ? '1px solid #1976D2' : '1px solid #ccc',
-                  background: nullActive ? '#E3F2FD' : '#fafafa',
-                  color: nullActive ? '#0D47A1' : '#444',
-                  borderRadius: '999px',
-                  padding: '0.25rem 0.5rem',
-                  fontSize: '0.7rem',
-                  cursor: 'pointer',
-                  textAlign: 'left'
-                }}
-                title={`Null values (${nullCount} ${metricLabels.short})`}
-              >
-                (Null) — {nullCount} {metricLabels.short}
-              </button>
-            </>
-          )
-        })()}
-        <div style={{ display: 'flex', gap: '0.25rem' }}>
-          <button
-            onClick={() => clearColumnFilter(tableName, columnName, cacheKey)}
-            style={{
-              border: 'none',
-              background: columnHasFilter ? '#1976D2' : '#eee',
-              color: columnHasFilter ? 'white' : '#555',
-              borderRadius: '4px',
-              padding: '0.25rem 0.5rem',
-              fontSize: '0.7rem',
-              cursor: columnHasFilter ? 'pointer' : 'default',
-              opacity: columnHasFilter ? 1 : 0.6,
-              flex: 1
-            }}
-            disabled={!columnHasFilter}
-          >
-            Reset
-          </button>
-          <button
-            onClick={toggleColumnNot}
-            style={{
-              border: 'none',
-              background: isNot ? '#333' : '#f0f0f0',
-              color: isNot ? 'white' : '#555',
-              borderRadius: '4px',
-              padding: '0.25rem 0.5rem',
-              fontSize: '0.7rem',
-              cursor: columnHasFilter ? 'pointer' : 'default',
-              opacity: columnHasFilter ? 1 : 0.6,
-              fontWeight: isNot ? 'bold' : 'normal'
-            }}
-            disabled={!columnHasFilter}
-            title={isNot ? 'Remove NOT' : 'Add NOT'}
-          >
-            ¬
-          </button>
-        </div>
-        {bins.length > 0 && (
-          <div style={{ borderTop: '1px solid #eee', paddingTop: '0.25rem', display: 'flex', flexWrap: 'wrap', gap: '0.25rem' }}>
-            {bins.map((bin, index) => {
-              const active = isRangeFiltered(tableName, columnName, bin.bin_start, bin.bin_end, cacheKey)
-              const label = `${formatRangeValue(bin.bin_start)} – ${formatRangeValue(bin.bin_end)}`
-              return (
-                <button
-                  key={`${tableName}-${columnName}-bin-${index}`}
-                  onMouseDown={event => event.preventDefault()}
-                  onClick={() => toggleRangeFilter(tableName, columnName, bin.bin_start, bin.bin_end, cacheKey)}
-                  style={{
-                    border: active ? '1px solid #1976D2' : '1px solid #ccc',
-                    background: active ? '#E3F2FD' : '#fafafa',
-                    color: active ? '#0D47A1' : '#444',
-                    borderRadius: '999px',
-                    padding: '0.25rem 0.5rem',
-                    fontSize: '0.7rem',
-                    cursor: 'pointer'
-                  }}
-                  title={`${label} (${bin.count} ${metricLabels.short})`}
-                >
-                  {label}
-                </button>
-              )
-            })}
-          </div>
-        )}
-        {customRanges.length > 0 && (
-          <div style={{ borderTop: '1px solid #eee', paddingTop: '0.25rem', display: 'flex', flexWrap: 'wrap', gap: '0.25rem' }}>
-            {customRanges.map((range, index) => {
-              const label = `${formatRangeValue(range.start)} – ${formatRangeValue(range.end)}`
-              return (
-                <button
-                  key={`${tableName}-${columnName}-custom-${index}`}
-                  onMouseDown={event => event.preventDefault()}
-                  onClick={() => updateColumnRanges(tableName, columnName, prev => prev.filter(r => !rangesEqual(r, range)), cacheKey)}
-                  style={{
-                    border: '1px solid #1976D2',
-                    background: '#E3F2FD',
-                    color: '#0D47A1',
-                    borderRadius: '999px',
-                    padding: '0.25rem 0.5rem',
-                    fontSize: '0.7rem',
-                    cursor: 'pointer'
-                  }}
-                  title={`Remove ${label}`}
-                >
-                  {label} ×
-                </button>
-              )
-            })}
-          </div>
-        )}
-        <div style={{ borderTop: '1px solid #eee', paddingTop: '0.35rem', display: 'flex', flexDirection: 'column', gap: '0.35rem' }}>
-          <div style={{ display: 'flex', gap: '0.35rem' }}>
-            <label style={{ fontSize: '0.7rem', color: '#555', flex: 1 }}>
-              From
-              <input
-                type="number"
-                value={range.min}
-                onChange={(event) => handleCustomRangeChange(key, 'min', event.target.value)}
-                placeholder={stats?.min !== null && stats?.min !== undefined ? String(stats.min) : ''}
-                style={{ width: '100%', padding: '0.2rem 0.3rem', marginTop: '0.15rem' }}
-              />
-            </label>
-            <label style={{ fontSize: '0.7rem', color: '#555', flex: 1 }}>
-              To
-              <input
-                type="number"
-                value={range.max}
-                onChange={(event) => handleCustomRangeChange(key, 'max', event.target.value)}
-                placeholder={stats?.max !== null && stats?.max !== undefined ? String(stats.max) : ''}
-                style={{ width: '100%', padding: '0.2rem 0.3rem', marginTop: '0.15rem' }}
-              />
-            </label>
-          </div>
-          <button
-            onClick={() => applyCustomRange(tableName, columnName, cacheKey)}
-            style={{
-              border: 'none',
-              background: hasValidRange ? '#1976D2' : '#ccc',
-              color: 'white',
-              borderRadius: '4px',
-              padding: '0.3rem 0.5rem',
-              fontSize: '0.75rem',
-              cursor: hasValidRange ? 'pointer' : 'default'
-            }}
-            disabled={!hasValidRange}
-          >
-            Apply
-          </button>
-        </div>
-      </div>
-    )
-  }
-
   const toggleFilter = (column: string, value: string | number, tableName?: string, countByKey?: string) => {
     const filterValue = normalizeFilterValue(value)
     const resolvedCountKey = countByKey ?? ROW_COUNT_KEY
@@ -2134,1542 +1708,54 @@ function DatasetExplorer() {
 
   const getCountByValueForTable = (tableName: string) => getCountByCacheKey(tableName)
 
-  const renderTableView = (
-    title: string,
-    tableName: string,
-    field: string,
-    tableColor?: string,
-    aggregationOverride?: ColumnAggregation,
-    cacheKeyOverride?: string,
-    countIndicatorOverride?: React.ReactNode
-  ) => {
-    const cacheKey = cacheKeyOverride ?? getEffectiveCacheKeyForChart(tableName, field)
-    const aggregation =
-      aggregationOverride && (!cacheKeyOverride || cacheKeyOverride === cacheKey)
-        ? aggregationOverride
-        : getAggregation(tableName, field, cacheKey)
-
-    const metadata = getColumnMetadata(tableName, field)
-    const tableDisplayName = getTableDisplayNameByName(tableName) || tableName
-    const tooltipParts = [
-      metadata?.display_name || title,
-      `ID: ${field}`,
-      metadata?.description || '',
-      `Table: ${tableDisplayName}`
-    ]
-    if (aggregation) {
-      const pathLabel = formatMetricPath(aggregation)
-      if (pathLabel) {
-        tooltipParts.push(pathLabel)
-      }
-    }
-    const tooltipText = tooltipParts.filter(Boolean).join('\n')
-
-    const menuOpen =
-      activeFilterMenu?.tableName === tableName &&
-      activeFilterMenu.columnName === field &&
-      activeFilterMenu.countKey === cacheKey
-    const columnActive = hasColumnFilter(field, cacheKey)
-
-    const actionButtons = (
-      <>
-        <button
-          type="button"
-          onClick={event => {
-            event.stopPropagation()
-            toggleViewPreference(tableName, field)
-          }}
-          style={{
-            border: 'none',
-            background: '#f0f0f0',
-            color: '#333',
-            borderRadius: '50%',
-            width: '20px',
-            height: '20px',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: '0.7rem',
-            cursor: 'pointer',
-            lineHeight: 1
-          }}
-          title="Switch to chart view"
-        >
-          ◐
-        </button>
-        <button
-          type="button"
-          onClick={event => {
-            event.stopPropagation()
-            toggleDashboard(tableName, field)
-          }}
-          style={{
-            border: 'none',
-            background: isOnDashboard(tableName, field) ? '#4CAF50' : '#f0f0f0',
-            color: isOnDashboard(tableName, field) ? 'white' : '#333',
-            borderRadius: '50%',
-            width: '20px',
-            height: '20px',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: '0.7rem',
-            cursor: 'pointer',
-            lineHeight: 1
-          }}
-          title={isOnDashboard(tableName, field) ? 'Remove from dashboard' : 'Add to dashboard'}
-        >
-          {isOnDashboard(tableName, field) ? '✓' : '+'}
-        </button>
-        <button
-          type="button"
-          onClick={event => {
-            event.stopPropagation()
-            setActiveFilterMenu(prev =>
-              prev && prev.tableName === tableName && prev.columnName === field && prev.countKey === cacheKey
-                ? null
-                : { tableName, columnName: field, countKey: cacheKey }
-            )
-          }}
-          style={{
-            border: 'none',
-            background: menuOpen || columnActive ? '#1976D2' : '#f0f0f0',
-            color: menuOpen || columnActive ? 'white' : '#333',
-            borderRadius: '50%',
-            width: '20px',
-            height: '20px',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: '0.75rem',
-            cursor: 'pointer',
-            lineHeight: 1
-          }}
-          title="Filter values"
-        >
-          ⚲
-        </button>
-      </>
-    )
-
-    const countIndicator = countIndicatorOverride ?? renderTableCountIndicator(tableName, field, cacheKey)
-
-    const containerStyle: React.CSSProperties = {
-      position: 'relative',
-      background: 'white',
-      padding: '0.5rem',
-      borderRadius: '8px',
-      boxShadow: '0 1px 3px rgba(0,0,0,0.1)',
-      width: '358px',
-      height: '358px',
-      boxSizing: 'border-box',
-      flexShrink: 0,
-      display: 'flex',
-      flexDirection: 'column',
-      border: tableColor ? `2px solid ${tableColor}20` : undefined
-    }
-
-    if (!aggregation || !aggregation.categories || aggregation.categories.length === 0) {
-      const message = aggregation ? 'No data for current filters' : 'Loading data…'
-      return (
-        <div style={containerStyle}>
-          {renderChartHeader({
-            title: metadata?.display_name || title,
-            tooltip: tooltipText,
-            countIndicator,
-            actions: actionButtons,
-            isListColumn: metadata?.is_list_column || false
-          })}
-          <div
-            style={{
-              flex: 1,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              color: '#999',
-              fontSize: '0.75rem',
-              textAlign: 'center',
-              padding: '0.5rem'
-            }}
-          >
-            {message}
-          </div>
-        </div>
-      )
-    }
-
-    const baselineAggregation = getBaselineAggregation(tableName, field)
-    const categoriesForMenu =
-      metricsMatch(baselineAggregation, aggregation) && baselineAggregation?.categories?.length
-        ? baselineAggregation.categories
-        : aggregation.categories
-
-    const totalRows = aggregation.total_rows ?? aggregation.categories.reduce((sum, cat) => sum + cat.count, 0)
-
-    const tableData = aggregation.categories.map(cat => ({
-      category: cat.display_value ?? (cat.value === '' ? '(Empty)' : String(cat.value)),
-      rawValue: cat.value,
-      count: cat.count,
-      percentage: totalRows > 0 ? (cat.count / totalRows) * 100 : 0
-    }))
-
-    const sortedData = [...tableData].sort((a, b) => b.count - a.count)
-    const showLimit = 100
-    const visibleData = sortedData.slice(0, showLimit)
-    const hasMore = sortedData.length > showLimit
-
-    return (
-      <div style={containerStyle}>
-        {renderChartHeader({
-          title: metadata?.display_name || title,
-          tooltip: tooltipText,
-          countIndicator,
-          actions: actionButtons,
-          isListColumn: metadata?.is_list_column || false
-        })}
-        <div
-          style={{
-            overflowY: 'auto',
-            flex: 1,
-            minHeight: 0,
-            fontSize: '0.75rem'
-          }}
-        >
-          <table
-            style={{
-              width: '100%',
-              borderCollapse: 'collapse',
-              fontSize: '0.75rem'
-            }}
-          >
-            <thead
-              style={{
-                position: 'sticky',
-                top: 0,
-                background: '#f5f5f5',
-                borderBottom: '2px solid #ddd'
-              }}
-            >
-              <tr>
-                <th
-                  style={{
-                    padding: '0.4rem 0.5rem',
-                    textAlign: 'left',
-                    fontWeight: 600
-                  }}
-                >
-                  Category
-                </th>
-                <th
-                  style={{
-                    padding: '0.4rem 0.5rem',
-                    textAlign: 'right',
-                    fontWeight: 600
-                  }}
-                >
-                  Count ↓
-                </th>
-                <th
-                  style={{
-                    padding: '0.4rem 0.5rem',
-                    textAlign: 'right',
-                    fontWeight: 600
-                  }}
-                >
-                  %
-                </th>
-              </tr>
-            </thead>
-            <tbody>
-              {visibleData.map((row, idx) => {
-                const rawValue = normalizeFilterValue(row.rawValue)
-                const isFiltered = isValueFiltered(field, rawValue, cacheKey)
-                return (
-                  <tr
-                    key={idx}
-                    onClick={() => toggleFilter(field, rawValue, tableName, cacheKey)}
-                    style={{
-                      cursor: 'pointer',
-                      background: isFiltered ? '#E3F2FD' : idx % 2 === 0 ? 'white' : '#fafafa',
-                      borderLeft: isFiltered ? '3px solid #1976D2' : '3px solid transparent'
-                    }}
-                    onMouseEnter={e => {
-                      if (!isFiltered) e.currentTarget.style.background = '#f0f0f0'
-                    }}
-                    onMouseLeave={e => {
-                      if (!isFiltered) e.currentTarget.style.background = idx % 2 === 0 ? 'white' : '#fafafa'
-                    }}
-                  >
-                    <td
-                      style={{
-                        padding: '0.4rem 0.5rem',
-                        overflow: 'hidden',
-                        textOverflow: 'ellipsis',
-                        whiteSpace: 'nowrap',
-                        maxWidth: '180px'
-                      }}
-                    >
-                      {row.category}
-                    </td>
-                    <td style={{ padding: '0.4rem 0.5rem', textAlign: 'right' }}>
-                      {row.count.toLocaleString()}
-                    </td>
-                    <td style={{ padding: '0.4rem 0.5rem', textAlign: 'right' }}>
-                      {row.percentage.toFixed(1)}%
-                    </td>
-                  </tr>
-                )
-              })}
-            </tbody>
-          </table>
-          {hasMore && (
-            <div style={{ textAlign: 'center', padding: '0.5rem', fontSize: '0.7rem', color: '#666' }}>
-              Showing first {showLimit} of {sortedData.length} categories
-            </div>
-          )}
-        </div>
-        {renderFilterMenu(tableName, field, categoriesForMenu, cacheKey)}
-      </div>
-    )
-  }
-
-  const renderPieChart = (
-    title: string,
-    tableName: string,
-    field: string,
-    tableColor?: string,
-    aggregationOverride?: ColumnAggregation,
-    cacheKeyOverride?: string,
-    countIndicatorOverride?: React.ReactNode
-  ) => {
-    const cacheKey = cacheKeyOverride ?? getEffectiveCacheKeyForChart(tableName, field)
-    const aggregation =
-      aggregationOverride && (!cacheKeyOverride || cacheKeyOverride === cacheKey)
-        ? aggregationOverride
-        : getAggregation(tableName, field, cacheKey)
-
-    const metadata = getColumnMetadata(tableName, field)
-    const tableDisplayName = getTableDisplayNameByName(tableName) || tableName
-    const tooltipParts = [
-      metadata?.display_name || title,
-      `ID: ${field}`,
-      metadata?.description || '',
-      `Table: ${tableDisplayName}`
-    ]
-    if (aggregation) {
-      const pathLabel = formatMetricPath(aggregation)
-      if (pathLabel) tooltipParts.push(pathLabel)
-    }
-    const tooltipText = tooltipParts.filter(Boolean).join('\n')
-
-    const menuOpen =
-      activeFilterMenu?.tableName === tableName &&
-      activeFilterMenu.columnName === field &&
-      activeFilterMenu.countKey === cacheKey
-    const columnActive = hasColumnFilter(field, cacheKey)
-    const actionButtons = (
-      <>
-        <button
-          type="button"
-          onClick={event => {
-            event.stopPropagation()
-            toggleViewPreference(tableName, field)
-          }}
-          style={{
-            border: 'none',
-            background: '#f0f0f0',
-            color: '#333',
-            borderRadius: '50%',
-            width: '20px',
-            height: '20px',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: '0.7rem',
-            cursor: 'pointer',
-            lineHeight: 1
-          }}
-          title="Switch to table view"
-        >
-          ⊞
-        </button>
-        <button
-          type="button"
-          onClick={event => {
-            event.stopPropagation()
-            toggleDashboard(tableName, field)
-          }}
-          style={{
-            border: 'none',
-            background: isOnDashboard(tableName, field) ? '#4CAF50' : '#f0f0f0',
-            color: isOnDashboard(tableName, field) ? 'white' : '#333',
-            borderRadius: '50%',
-            width: '20px',
-            height: '20px',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: '0.7rem',
-            cursor: 'pointer',
-            lineHeight: 1
-          }}
-          title={isOnDashboard(tableName, field) ? 'Remove from dashboard' : 'Add to dashboard'}
-        >
-          {isOnDashboard(tableName, field) ? '✓' : '+'}
-        </button>
-        <button
-          type="button"
-          onClick={event => {
-            event.stopPropagation()
-            setActiveFilterMenu(prev =>
-              prev && prev.tableName === tableName && prev.columnName === field && prev.countKey === cacheKey
-                ? null
-                : { tableName, columnName: field, countKey: cacheKey }
-            )
-          }}
-          style={{
-            border: 'none',
-            background: menuOpen || columnActive ? '#1976D2' : '#f0f0f0',
-            color: menuOpen || columnActive ? 'white' : '#333',
-            borderRadius: '50%',
-            width: '20px',
-            height: '20px',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: '0.75rem',
-            cursor: 'pointer',
-            lineHeight: 1
-          }}
-          title="Filter values"
-        >
-          ⚲
-        </button>
-      </>
-    )
-
-    const containerStyle: React.CSSProperties = {
-      position: 'relative',
-      background: 'white',
-      padding: '0.5rem',
-      borderRadius: '8px',
-      boxShadow: '0 1px 3px rgba(0,0,0,0.1)',
-      width: '175px',
-      minHeight: '175px',
-      boxSizing: 'border-box',
-      flexShrink: 0,
-      display: 'flex',
-      flexDirection: 'column',
-      border: tableColor ? `2px solid ${tableColor}20` : undefined
-    }
-
-    const countIndicator = countIndicatorOverride ?? renderTableCountIndicator(tableName, field, cacheKey)
-
-    if (!aggregation || !aggregation.categories || aggregation.categories.length === 0) {
-      const message = aggregation ? 'No data for current filters' : 'Loading data…'
-      return (
-        <div style={containerStyle}>
-          {renderChartHeader({
-            title: metadata?.display_name || title,
-            tooltip: tooltipText,
-            countIndicator,
-            actions: actionButtons,
-            isListColumn: metadata?.is_list_column || false
-          })}
-          <div
-            style={{
-              flex: 1,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              color: '#999',
-              fontSize: '0.75rem',
-              textAlign: 'center',
-              padding: '0.5rem'
-            }}
-          >
-            {message}
-          </div>
-        </div>
-      )
-    }
-
-    const metricLabels = getMetricLabels(aggregation)
-    const pathLabel = formatMetricPath(aggregation)
-    const labels = aggregation.categories.map(c => c.display_value ?? (c.value === '' ? '(Empty)' : String(c.value)))
-    const values = aggregation.categories.map(c => c.count)
-    const filterValues = aggregation.categories.map(c => normalizeFilterValue(c.value))
-    const shouldShowPiePercentages = showPercentageLabels
-
-    const baselineAggregation = getBaselineAggregation(tableName, field)
-    const categoriesForMenu =
-      metricsMatch(baselineAggregation, aggregation) && baselineAggregation?.categories?.length
-        ? baselineAggregation.categories
-        : aggregation.categories
-
-    const totalCount = aggregation.total_rows ?? values.reduce((sum, val) => sum + val, 0)
-    const percentTexts = values.map(val =>
-      totalCount > 0 ? `${((val / totalCount) * 100).toFixed(1)}%` : '0%'
-    )
-    const countTexts = values.map(val => val.toLocaleString())
-
-    return (
-      <div style={containerStyle}>
-        {renderChartHeader({
-          title: metadata?.display_name || title,
-          tooltip: tooltipText,
-          countIndicator,
-          actions: actionButtons,
-          isListColumn: metadata?.is_list_column || false
-        })}
-        <Plot
-          data={[{
-            type: 'pie',
-            labels,
-            values,
-            textposition: 'inside',
-            insidetextorientation: 'radial',
-            marker: {
-              colors: filterValues.map(value =>
-                isValueFiltered(field, value, cacheKey) ? '#1976D2' : undefined
-              ),
-              line: {
-                color: filterValues.map(value =>
-                  isValueFiltered(field, value, cacheKey) ? '#000' : undefined
-                ),
-                width: filterValues.map(value =>
-                  isValueFiltered(field, value, cacheKey) ? 2 : 0
-                )
-              }
-            },
-            textfont: { size: 9 },
-            hovertemplate: `${['%{label}', `Count (${metricLabels.short}): %{value}`, 'Percent of total: %{customdata}']
-              .concat(pathLabel ? [pathLabel] : [])
-              .join('<br>')}<extra></extra>`,
-            customdata: percentTexts,
-            textinfo: shouldShowPiePercentages ? 'label+text' : 'label+value',
-            text: shouldShowPiePercentages ? percentTexts : countTexts
-          }]}
-          layout={{
-            height: 135,
-            margin: { t: 5, b: 5, l: 5, r: 5 },
-            showlegend: false,
-            paper_bgcolor: 'transparent',
-            plot_bgcolor: 'transparent',
-            dragmode: false
-          }}
-          config={{
-            displayModeBar: false,
-            responsive: true,
-            staticPlot: false
-          }}
-          style={{ width: '165px', height: '135px', cursor: 'pointer' }}
-          onClick={(event: PlotMouseEvent) => {
-            const point = event.points?.[0]
-            if (!point) return
-
-            const index = point.pointNumber ?? point.pointIndex
-            if (typeof index === 'number' && index >= 0 && index < filterValues.length) {
-              const clickedValue = filterValues[index]
-              toggleFilter(field, clickedValue, tableName, cacheKey)
-            }
-          }}
-        />
-        {renderFilterMenu(tableName, field, categoriesForMenu, cacheKey)}
-      </div>
-    )
-  }
-
-  const renderBarChart = (
-    title: string,
-    tableName: string,
-    field: string,
-    tableColor?: string,
-    aggregationOverride?: ColumnAggregation,
-    cacheKeyOverride?: string,
-    countIndicatorOverride?: React.ReactNode
-  ) => {
-    const cacheKey = cacheKeyOverride ?? getEffectiveCacheKeyForChart(tableName, field)
-    const aggregation =
-      aggregationOverride && (!cacheKeyOverride || cacheKeyOverride === cacheKey)
-        ? aggregationOverride
-        : getAggregation(tableName, field, cacheKey)
-
-    const metadata = getColumnMetadata(tableName, field)
-    const tableDisplayName = getTableDisplayNameByName(tableName) || tableName
-    const tooltipParts = [
-      metadata?.display_name || title,
-      `ID: ${field}`,
-      metadata?.description || '',
-      `Table: ${tableDisplayName}`
-    ]
-    if (aggregation) {
-      const pathLabel = formatMetricPath(aggregation)
-      if (pathLabel) tooltipParts.push(pathLabel)
-    }
-    const tooltipText = tooltipParts.filter(Boolean).join('\n')
-
-    const menuOpen =
-      activeFilterMenu?.tableName === tableName &&
-      activeFilterMenu.columnName === field &&
-      activeFilterMenu.countKey === cacheKey
-    const columnActive = hasColumnFilter(field, cacheKey)
-
-    const actionButtons = (
-      <>
-        <button
-          type="button"
-          onClick={event => {
-            event.stopPropagation()
-            toggleViewPreference(tableName, field)
-          }}
-          style={{
-            border: 'none',
-            background: '#f0f0f0',
-            color: '#333',
-            borderRadius: '50%',
-            width: '20px',
-            height: '20px',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: '0.7rem',
-            cursor: 'pointer',
-            lineHeight: 1
-          }}
-          title="Switch to table view"
-        >
-          ⊞
-        </button>
-        <button
-          type="button"
-          onClick={event => {
-            event.stopPropagation()
-            toggleDashboard(tableName, field)
-          }}
-          style={{
-            border: 'none',
-            background: isOnDashboard(tableName, field) ? '#4CAF50' : '#f0f0f0',
-            color: isOnDashboard(tableName, field) ? 'white' : '#333',
-            borderRadius: '50%',
-            width: '20px',
-            height: '20px',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: '0.7rem',
-            cursor: 'pointer',
-            lineHeight: 1
-          }}
-          title={isOnDashboard(tableName, field) ? 'Remove from dashboard' : 'Add to dashboard'}
-        >
-          {isOnDashboard(tableName, field) ? '✓' : '+'}
-        </button>
-        <button
-          type="button"
-          onClick={event => {
-            event.stopPropagation()
-            setActiveFilterMenu(prev =>
-              prev && prev.tableName === tableName && prev.columnName === field && prev.countKey === cacheKey
-                ? null
-                : { tableName, columnName: field, countKey: cacheKey }
-            )
-          }}
-          style={{
-            border: 'none',
-            background: menuOpen || columnActive ? '#1976D2' : '#f0f0f0',
-            color: menuOpen || columnActive ? 'white' : '#333',
-            borderRadius: '50%',
-            width: '20px',
-            height: '20px',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: '0.75rem',
-            cursor: 'pointer',
-            lineHeight: 1
-          }}
-          title="Filter values"
-        >
-          ⚲
-        </button>
-      </>
-    )
-
-    const containerStyle: React.CSSProperties = {
-      position: 'relative',
-      background: 'white',
-      padding: '0.5rem',
-      borderRadius: '8px',
-      boxShadow: '0 1px 3px rgba(0,0,0,0.1)',
-      width: '358px',
-      minHeight: '175px',
-      boxSizing: 'border-box',
-      flexShrink: 0,
-      display: 'flex',
-      flexDirection: 'column',
-      border: tableColor ? `2px solid ${tableColor}20` : undefined
-    }
-
-    const countIndicator = countIndicatorOverride ?? renderTableCountIndicator(tableName, field, cacheKey)
-
-    if (!aggregation || !aggregation.categories || aggregation.categories.length === 0) {
-      const message = aggregation ? 'No data for current filters' : 'Loading data…'
-      return (
-        <div style={containerStyle}>
-          {renderChartHeader({
-            title: metadata?.display_name || title,
-            tooltip: tooltipText,
-            countIndicator,
-            actions: actionButtons,
-            isListColumn: metadata?.is_list_column || false
-          })}
-          <div
-            style={{
-              flex: 1,
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              color: '#999',
-              fontSize: '0.75rem',
-              textAlign: 'center',
-              padding: '0.5rem'
-            }}
-          >
-            {message}
-          </div>
-        </div>
-      )
-    }
-
-    const metricLabels = getMetricLabels(aggregation)
-    const pathLabel = formatMetricPath(aggregation)
-    const labels = aggregation.categories.map(c => c.display_value ?? (c.value === '' ? '(Empty)' : String(c.value)))
-    const values = aggregation.categories.map(c => c.count)
-    const filterValues = aggregation.categories.map(c => normalizeFilterValue(c.value))
-    const totalCount = aggregation.total_rows ?? values.reduce((sum, val) => sum + val, 0)
-    const percentTexts = values.map(val =>
-      totalCount > 0 ? `${((val / totalCount) * 100).toFixed(1)}%` : '0%'
-    )
-
-    const baselineAggregation = getBaselineAggregation(tableName, field)
-    const categoriesForMenu =
-      metricsMatch(baselineAggregation, aggregation) && baselineAggregation?.categories?.length
-        ? baselineAggregation.categories
-        : aggregation.categories
-
-    return (
-      <div style={containerStyle}>
-        {renderChartHeader({
-          title: metadata?.display_name || title,
-          tooltip: tooltipText,
-          countIndicator,
-          actions: actionButtons,
-          isListColumn: metadata?.is_list_column || false
-        })}
-        <Plot
-          data={[{
-            type: 'bar',
-            x: labels,
-            y: values,
-            marker: {
-              color: filterValues.map(value =>
-                isValueFiltered(field, value, cacheKey) ? '#1976D2' : '#2196F3'
-              ),
-              line: {
-                color: filterValues.map(value =>
-                  isValueFiltered(field, value, cacheKey) ? '#000' : undefined
-                ),
-                width: filterValues.map(value =>
-                  isValueFiltered(field, value, cacheKey) ? 2 : 0
-                )
-              }
-            },
-            hovertemplate: `${['%{x}', `Count (${metricLabels.short}): %{y}`, 'Percent of total: %{text}']
-              .concat(pathLabel ? [pathLabel] : [])
-              .join('<br>')}<extra></extra>`,
-            text: percentTexts,
-            textposition: 'auto'
-          }]}
-          layout={{
-            height: 135,
-            margin: { t: 5, b: 40, l: 30, r: 5 },
-            xaxis: { tickangle: -45, automargin: true, tickfont: { size: 9 } },
-            yaxis: { title: metricLabels.long, automargin: true, tickfont: { size: 9 }, titlefont: { size: 10 } },
-            paper_bgcolor: 'transparent',
-            plot_bgcolor: 'transparent',
-            dragmode: 'select',
-            selectdirection: 'h'
-          }}
-          config={{
-            displayModeBar: false,
-            responsive: true,
-            staticPlot: false,
-            scrollZoom: false
-          }}
-          style={{ width: '348px', height: '135px', cursor: 'pointer' }}
-          onClick={(event: PlotMouseEvent) => {
-            const point = event.points?.[0]
-            if (!point) return
-
-            const pointIndex = point.pointIndex
-            if (typeof pointIndex === 'number' && pointIndex >= 0 && pointIndex < filterValues.length) {
-              const clickedValue = filterValues[pointIndex]
-              toggleFilter(field, clickedValue, tableName, cacheKey)
-            }
-          }}
-          onSelected={(event: PlotSelectionEvent) => {
-            if (!event?.points || event.points.length === 0) return
-            const selectedValues = event.points
-              .map(p => p.pointIndex)
-              .filter((idx): idx is number => typeof idx === 'number' && idx >= 0 && idx < filterValues.length)
-              .map(idx => filterValues[idx])
-
-            if (selectedValues.length > 0) {
-              setFilters(prev => [
-                ...removeColumnFilters(prev, field, cacheKey),
-                { column: field, operator: 'in', value: selectedValues, tableName, countByKey: cacheKey }
-              ])
-            }
-          }}
-        />
-        {renderFilterMenu(tableName, field, categoriesForMenu, cacheKey)}
-      </div>
-    )
-  }
-
-  const renderHistogram = (
-    title: string,
-    tableName: string,
-    field: string,
-    tableColor?: string,
-    aggregationOverride?: ColumnAggregation,
-    cacheKeyOverride?: string,
-    countIndicatorOverride?: React.ReactNode,
-    _extraActions?: React.ReactNode
-  ) => {
-    const cacheKey = cacheKeyOverride ?? getEffectiveCacheKeyForChart(tableName, field)
-    const aggregation =
-      aggregationOverride && (!cacheKeyOverride || cacheKeyOverride === cacheKey)
-        ? aggregationOverride
-        : getAggregation(tableName, field, cacheKey)
-    if (!aggregation?.numeric_stats) return null
-
-    const rawHistogram = aggregation.histogram ?? []
-    if (rawHistogram.length === 0) return null
-
-    const metricLabels = getMetricLabels(aggregation)
-    const pathLabel = formatMetricPath(aggregation)
-    const metadata = getColumnMetadata(tableName, field)
-    const tableDisplayName = getTableDisplayNameByName(tableName) || tableName
-
-    const statsText = [
-      `Mean: ${aggregation.numeric_stats.mean !== null ? aggregation.numeric_stats.mean.toFixed(2) : 'N/A'}`,
-      `Median: ${aggregation.numeric_stats.median !== null ? aggregation.numeric_stats.median.toFixed(2) : 'N/A'}`,
-      `Range: [${aggregation.numeric_stats.min !== null ? aggregation.numeric_stats.min.toFixed(2) : 'N/A'}, ${aggregation.numeric_stats.max !== null ? aggregation.numeric_stats.max.toFixed(2) : 'N/A'}]`
-    ].join(' | ')
-
-    const tooltipParts = [
-      metadata?.display_name || title,
-      `ID: ${field}`,
-      metadata?.description || '',
-      `Table: ${tableDisplayName}`
-    ]
-    if (pathLabel) tooltipParts.push(pathLabel)
-    tooltipParts.push('', statsText)
-    const tooltipText = tooltipParts.filter(Boolean).join('\n')
-
-    const baselineAggregation = getBaselineAggregation(tableName, field)
-    const histogramMatches = metricsMatch(baselineAggregation, aggregation)
-    const menuHistogram = histogramMatches && baselineAggregation?.histogram?.length
-      ? baselineAggregation.histogram
-      : rawHistogram
-    const menuStats = histogramMatches && baselineAggregation?.numeric_stats
-      ? baselineAggregation.numeric_stats
-      : aggregation.numeric_stats
-
-    const menuOpen =
-      activeFilterMenu?.tableName === tableName &&
-      activeFilterMenu.columnName === field &&
-      activeFilterMenu.countKey === cacheKey
-    const columnActive = hasColumnFilter(field, cacheKey)
-
-    const actionButtons = (
-      <>
-        <button
-          type="button"
-          onClick={event => {
-            event.stopPropagation()
-            toggleDashboard(tableName, field)
-          }}
-          style={{
-            border: 'none',
-            background: isOnDashboard(tableName, field) ? '#4CAF50' : '#f0f0f0',
-            color: isOnDashboard(tableName, field) ? 'white' : '#333',
-            borderRadius: '50%',
-            width: '20px',
-            height: '20px',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: '0.7rem',
-            cursor: 'pointer',
-            lineHeight: 1
-          }}
-          title={isOnDashboard(tableName, field) ? 'Remove from dashboard' : 'Add to dashboard'}
-        >
-          {isOnDashboard(tableName, field) ? '✓' : '+'}
-        </button>
-        <button
-          type="button"
-          onClick={event => {
-            event.stopPropagation()
-            setActiveFilterMenu(prev =>
-              prev && prev.tableName === tableName && prev.columnName === field && prev.countKey === cacheKey
-                ? null
-                : { tableName, columnName: field, countKey: cacheKey }
-            )
-          }}
-          style={{
-            border: 'none',
-            background: menuOpen || columnActive ? '#1976D2' : '#f0f0f0',
-            color: menuOpen || columnActive ? 'white' : '#333',
-            borderRadius: '50%',
-            width: '20px',
-            height: '20px',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: '0.75rem',
-            cursor: 'pointer',
-            lineHeight: 1
-          }}
-          title="Filter values"
-        >
-          ⚲
-        </button>
-      </>
-    )
-
-    const displayHistogram = getDisplayHistogram(menuHistogram, menuStats)
-    const binsForPlot = displayHistogram.length > 0 ? displayHistogram : menuHistogram
-
-    const xValues = binsForPlot.map(bin => (bin.bin_start + bin.bin_end) / 2)
-    const yValues = binsForPlot.map(baselineBin => {
-      let totalCount = 0
-      rawHistogram.forEach(filteredBin => {
-        const overlapStart = Math.max(baselineBin.bin_start, filteredBin.bin_start)
-        const overlapEnd = Math.min(baselineBin.bin_end, filteredBin.bin_end)
-        if (overlapStart < overlapEnd) {
-          const filteredBinWidth = filteredBin.bin_end - filteredBin.bin_start
-          const overlapWidth = overlapEnd - overlapStart
-          const overlapFraction = overlapWidth / filteredBinWidth
-          totalCount += filteredBin.count * overlapFraction
-        }
-      })
-      return totalCount
-    })
-    const totalMetricCount = aggregation.total_rows ?? rawHistogram.reduce((sum, bin) => sum + bin.count, 0)
-    const sumY = yValues.reduce((sum, val) => sum + val, 0)
-    const scalingFactor = sumY > 0 && totalMetricCount > 0 ? totalMetricCount / sumY : 1
-    const adjustedYValues = scalingFactor < 1 ? yValues.map(val => val * scalingFactor) : yValues
-    const roundedYValues = adjustedYValues.map(val => Math.max(0, val))
-    const binWidth = binsForPlot[0] ? binsForPlot[0].bin_end - binsForPlot[0].bin_start : 1
-    const totalCount = totalMetricCount > 0 ? totalMetricCount : roundedYValues.reduce((sum, val) => sum + val, 0)
-    const percentTexts = roundedYValues.map(val =>
-      totalCount > 0 ? `${((val / totalCount) * 100).toFixed(1)}%` : '0%'
-    )
-    const countTexts = roundedYValues.map(val =>
-      val >= 1000 ? Math.round(val).toLocaleString() : Math.round(val).toString()
-    )
-
-    const containerStyle: React.CSSProperties = {
-      position: 'relative',
-      background: 'white',
-      padding: '0.5rem',
-      borderRadius: '8px',
-      boxShadow: '0 1px 3px rgba(0,0,0,0.1)',
-      width: '358px',
-      minHeight: '175px',
-      boxSizing: 'border-box',
-      flexShrink: 0,
-      display: 'flex',
-      flexDirection: 'column',
-      border: tableColor ? `2px solid ${tableColor}20` : undefined
-    }
-
-    const countIndicator = countIndicatorOverride ?? renderTableCountIndicator(tableName, field, cacheKey)
-
-    return (
-      <div style={containerStyle}>
-        {renderChartHeader({
-          title: metadata?.display_name || title,
-          tooltip: tooltipText,
-          countIndicator,
-          actions: actionButtons,
-          isListColumn: metadata?.is_list_column || false
-        })}
-        <Plot
-          data={[{
-            type: 'bar',
-            x: xValues,
-            y: roundedYValues,
-            width: binWidth * 0.9,
-            marker: {
-              color: binsForPlot.map(bin =>
-                isRangeFiltered(tableName, field, bin.bin_start, bin.bin_end, cacheKey) ? '#2E7D32' : '#4CAF50'
-              ),
-              line: {
-                color: binsForPlot.map(bin =>
-                  isRangeFiltered(tableName, field, bin.bin_start, bin.bin_end, cacheKey) ? '#000' : undefined
-                ),
-                width: binsForPlot.map(bin =>
-                  isRangeFiltered(tableName, field, bin.bin_start, bin.bin_end, cacheKey) ? 2 : 0
-                )
-              }
-            },
-            hovertemplate: `${[
-              'Range: [%{customdata[0]:.2f}, %{customdata[1]:.2f}]',
-              `Count (${metricLabels.short}): %{y}`,
-              'Percent of total: %{customdata[2]}'
-            ]
-              .concat(pathLabel ? [pathLabel] : [])
-              .join('<br>')}<extra></extra>`,
-            customdata: binsForPlot.map((bin, idx) => [bin.bin_start, bin.bin_end, percentTexts[idx]]),
-            text: showPercentageLabels ? percentTexts : countTexts,
-            textposition: 'auto'
-          }]}
-          layout={{
-            height: 135,
-            margin: { t: 5, b: 30, l: 30, r: 5 },
-            xaxis: { title: field, automargin: true, tickfont: { size: 9 }, titlefont: { size: 10 } },
-            yaxis: { title: metricLabels.long, automargin: true, tickfont: { size: 9 }, titlefont: { size: 10 } },
-            paper_bgcolor: 'transparent',
-            plot_bgcolor: 'transparent',
-            bargap: 0.1,
-            dragmode: 'select',
-            selectdirection: 'h'
-          }}
-          config={{
-            displayModeBar: false,
-            responsive: true,
-            staticPlot: false,
-            scrollZoom: false
-          }}
-          style={{ width: '348px', height: '135px', cursor: 'pointer' }}
-          onClick={(event: PlotMouseEvent) => {
-            const point = event.points?.[0]
-            if (!point) return
-            const pointIndex = point.pointIndex
-            if (typeof pointIndex === 'number' && pointIndex >= 0 && pointIndex < binsForPlot.length) {
-              const bin = binsForPlot[pointIndex]
-              toggleRangeFilter(tableName, field, bin.bin_start, bin.bin_end, cacheKey)
-            }
-          }}
-          onSelected={(event: PlotSelectionEvent) => {
-            const rangeX = event?.range?.x
-            if (!rangeX || rangeX.length < 2) return
-            const [minX, maxX] = rangeX
-            updateColumnRanges(tableName, field, prev => {
-              const nextRange = { start: minX, end: maxX }
-              const existingIndex = prev.findIndex(range => rangesEqual(range, nextRange))
-              if (existingIndex >= 0) return prev
-              return [...prev, nextRange]
-            }, cacheKey)
-          }}
-        />
-        {renderNumericFilterMenu(tableName, field, displayHistogram, menuStats, cacheKey)}
-      </div>
-    )
-  }
-
-  const renderSurvivalChart = (
-    title: string,
-    tableName: string,
-    field: string,
-    tableColor?: string,
-    aggregationOverride?: ColumnAggregation,
-    cacheKeyOverride?: string,
-    countIndicatorOverride?: React.ReactNode,
-    _extraActions?: React.ReactNode,
-    showHistogram: boolean = true
-  ) => {
-    const cacheKey = cacheKeyOverride ?? getEffectiveCacheKeyForChart(tableName, field)
-    const aggregation =
-      aggregationOverride && (!cacheKeyOverride || cacheKeyOverride === cacheKey)
-        ? aggregationOverride
-        : getAggregation(tableName, field, cacheKey)
-    const table = dataset?.tables.find(t => t.name === tableName)
-    if (!aggregation || !table) return null
-
-    const statusColumn = findSurvivalStatusColumn(tableName, field)
-    if (statusColumn) {
-      ensureSurvivalCurve(table, field, statusColumn, cacheKey)
-    }
-    const curve = statusColumn ? getSurvivalCurve(tableName, field, statusColumn, cacheKey) : undefined
-
-    const metadata = getColumnMetadata(tableName, field)
-    const tableDisplayName = getTableDisplayNameByName(tableName) || tableName
-    const tooltipParts = [
-      metadata?.display_name || title,
-      `ID: ${field}`,
-      metadata?.description || '',
-      `Table: ${tableDisplayName}`
-    ]
-    const pathLabel = formatMetricPath(aggregation)
-    if (pathLabel) tooltipParts.push(pathLabel)
-    const tooltipText = tooltipParts.filter(Boolean).join('\n')
-
-    const menuOpen =
-      activeFilterMenu?.tableName === tableName &&
-      activeFilterMenu.columnName === field &&
-      activeFilterMenu.countKey === cacheKey
-    const columnActive = hasColumnFilter(field, cacheKey)
-
-    const actionButtons = (
-      <>
-        <button
-          type="button"
-          onClick={event => {
-            event.stopPropagation()
-            toggleDashboard(tableName, field)
-          }}
-          style={{
-            border: 'none',
-            background: isOnDashboard(tableName, field) ? '#4CAF50' : '#f0f0f0',
-            color: isOnDashboard(tableName, field) ? 'white' : '#333',
-            borderRadius: '50%',
-            width: '20px',
-            height: '20px',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: '0.7rem',
-            cursor: 'pointer',
-            lineHeight: 1
-          }}
-          title={isOnDashboard(tableName, field) ? 'Remove from dashboard' : 'Add to dashboard'}
-        >
-          {isOnDashboard(tableName, field) ? '✓' : '+'}
-        </button>
-        <button
-          type="button"
-          onClick={event => {
-            event.stopPropagation()
-            setActiveFilterMenu(prev =>
-              prev && prev.tableName === tableName && prev.columnName === field && prev.countKey === cacheKey
-                ? null
-                : { tableName, columnName: field, countKey: cacheKey }
-            )
-          }}
-          style={{
-            border: 'none',
-            background: menuOpen || columnActive ? '#1976D2' : '#f0f0f0',
-            color: menuOpen || columnActive ? 'white' : '#333',
-            borderRadius: '50%',
-            width: '20px',
-            height: '20px',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: '0.75rem',
-            cursor: 'pointer',
-            lineHeight: 1
-          }}
-          title="Filter values"
-        >
-          ⚲
-        </button>
-      </>
-    )
-
-    const countIndicator = countIndicatorOverride ?? renderTableCountIndicator(tableName, field, cacheKey)
-    const displayHistogram = getDisplayHistogram(aggregation.histogram ?? [], aggregation.numeric_stats)
-    const binsForPlot = displayHistogram.length > 0 ? displayHistogram : aggregation.histogram || []
-    const histogramPlot =
-      binsForPlot.length > 0 ? (
-        <Plot
-          data={[{
-            type: 'bar',
-            x: binsForPlot.map(bin => (bin.bin_start + bin.bin_end) / 2),
-            y: binsForPlot.map(bin => bin.count),
-            width: binsForPlot.map(bin => bin.bin_end - bin.bin_start),
-            marker: { color: tableColor || '#2196F3', opacity: 0.7 }
-          }]}
-          layout={{
-            height: 180,
-            margin: { t: 20, b: 40, l: 50, r: 10 },
-            xaxis: { title: metadata?.display_name || title, tickfont: { size: 9 } },
-            yaxis: { title: getMetricLabels(aggregation).long, tickfont: { size: 9 } },
-            paper_bgcolor: 'transparent',
-            plot_bgcolor: 'transparent',
-            bargap: 0
-          }}
-          config={{
-            displayModeBar: false,
-            responsive: true,
-            staticPlot: false,
-            scrollZoom: false
-          }}
-          style={{ width: '100%', height: '180px' }}
-        />
-      ) : (
-        <div style={{ padding: '1rem', color: '#777', fontSize: '0.85rem' }}>
-          No histogram data available
-        </div>
-      )
-
-    const survivalPlot = statusColumn
-      ? curve && curve.length > 0 ? (
-        <Plot
-          data={[{
-            type: 'scatter',
-            mode: 'lines',
-            line: { shape: 'hv', color: tableColor || '#1976D2', width: 2 },
-            x: curve.map(p => p.time),
-            y: curve.map(p => p.survival),
-            customdata: curve.map(p => [p.atRisk, p.events, p.censored]),
-            hovertemplate: [
-              'Time: %{x}',
-              'Survival: %{y:.3f}',
-              'At risk: %{customdata[0]}',
-              'Events: %{customdata[1]}',
-              'Censored: %{customdata[2]}'
-            ].join('<br>') + '<extra></extra>'
-          }]}
-          layout={{
-            height: 260,
-            margin: { t: 20, b: 40, l: 50, r: 10 },
-            xaxis: { title: metadata?.display_name || title, tickfont: { size: 10 } },
-            yaxis: { title: 'Survival probability', range: [0, 1], tickfont: { size: 10 } },
-            paper_bgcolor: 'transparent',
-            plot_bgcolor: 'transparent',
-            hovermode: 'closest'
-          }}
-          config={{
-            displayModeBar: false,
-            responsive: true,
-            staticPlot: false,
-            scrollZoom: false
-          }}
-          style={{ width: '100%', height: '260px' }}
-        />
-      ) : (
-        <div style={{ padding: '1rem', color: '#777', fontSize: '0.85rem' }}>
-          {curve ? 'No survival data for current filters' : 'Loading survival curve…'}
-        </div>
-      )
-      : (
-        <div style={{ padding: '1rem', color: '#777', fontSize: '0.85rem' }}>
-          Add a survival status column to plot a Kaplan–Meier curve.
-        </div>
-      )
-
-    const containerStyle: React.CSSProperties = {
-      position: 'relative',
-      background: 'white',
-      padding: '0.5rem',
-      borderRadius: '8px',
-      boxShadow: '0 1px 3px rgba(0,0,0,0.1)',
-      width: '100%',
-      minHeight: showHistogram ? '420px' : '320px',
-      boxSizing: 'border-box',
-      flexShrink: 0,
-      display: 'flex',
-      flexDirection: 'column',
-      border: tableColor ? `2px solid ${tableColor}20` : undefined
-    }
-
-    const menuStats = aggregation.numeric_stats
-    const displayHistogramForMenu = getDisplayHistogram(aggregation.histogram ?? [], aggregation.numeric_stats)
-
-    return (
-      <div style={containerStyle}>
-        {renderChartHeader({
-          title: metadata?.display_name || title,
-          tooltip: tooltipText,
-          countIndicator,
-          actions: actionButtons,
-          isListColumn: metadata?.is_list_column || false
-        })}
-        <div style={{ display: 'grid', gridTemplateColumns: showHistogram ? '2fr 1fr' : '1fr', gap: '0.5rem', flex: 1 }}>
-          <div style={{ background: '#fafafa', borderRadius: '6px', padding: '0.35rem' }}>
-            {survivalPlot}
-          </div>
-          {showHistogram && (
-            <div style={{ background: '#fafafa', borderRadius: '6px', padding: '0.35rem' }}>
-              {histogramPlot}
-            </div>
-          )}
-        </div>
-        {renderNumericFilterMenu(tableName, field, displayHistogramForMenu, menuStats, cacheKey)}
-      </div>
-    )
-  }
-
-  const renderMapChart = (
-    title: string,
-    tableName: string,
-    field: string,
-    tableColor?: string,
-    aggregationOverride?: ColumnAggregation,
-    cacheKeyOverride?: string,
-    countIndicatorOverride?: React.ReactNode
-  ) => {
-    const cacheKey = cacheKeyOverride ?? getEffectiveCacheKeyForChart(tableName, field)
-    const aggregation =
-      aggregationOverride && (!cacheKeyOverride || cacheKeyOverride === cacheKey)
-        ? aggregationOverride
-        : getAggregation(tableName, field, cacheKey)
-
-    if (!aggregation?.categories) return null
-
-    const metadata = getColumnMetadata(tableName, field)
-    const tableDisplayName = getTableDisplayNameByName(tableName) || tableName
-    const tooltipParts = [
-      metadata?.display_name || title,
-      `ID: ${field}`,
-      metadata?.description || '',
-      `Table: ${tableDisplayName}`
-    ]
-    if (aggregation) {
-      const pathLabel = formatMetricPath(aggregation)
-      if (pathLabel) tooltipParts.push(pathLabel)
-    }
-    const tooltipText = tooltipParts.filter(Boolean).join('\n')
-
-    const menuOpen =
-      activeFilterMenu?.tableName === tableName &&
-      activeFilterMenu.columnName === field &&
-      activeFilterMenu.countKey === cacheKey
-    const columnActive = hasColumnFilter(field, cacheKey)
-
-    const actionButtons = (
-      <>
-        <button
-          type="button"
-          onClick={event => {
-            event.stopPropagation()
-            toggleDashboard(tableName, field)
-          }}
-          style={{
-            border: 'none',
-            background: isOnDashboard(tableName, field) ? '#4CAF50' : '#f0f0f0',
-            color: isOnDashboard(tableName, field) ? 'white' : '#333',
-            borderRadius: '50%',
-            width: '20px',
-            height: '20px',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: '0.7rem',
-            cursor: 'pointer',
-            lineHeight: 1
-          }}
-          title={isOnDashboard(tableName, field) ? 'Remove from dashboard' : 'Add to dashboard'}
-        >
-          {isOnDashboard(tableName, field) ? '✓' : '+'}
-        </button>
-        <button
-          type="button"
-          onClick={event => {
-            event.stopPropagation()
-            setActiveFilterMenu(prev =>
-              prev && prev.tableName === tableName && prev.columnName === field && prev.countKey === cacheKey
-                ? null
-                : { tableName, columnName: field, countKey: cacheKey }
-            )
-          }}
-          style={{
-            border: 'none',
-            background: menuOpen || columnActive ? '#1976D2' : '#f0f0f0',
-            color: menuOpen || columnActive ? 'white' : '#333',
-            borderRadius: '50%',
-            width: '20px',
-            height: '20px',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            fontSize: '0.7rem',
-            cursor: 'pointer',
-            lineHeight: 1
-          }}
-          title={columnActive ? 'Active filter' : 'Filter'}
-        >
-          ≡
-        </button>
-      </>
-    )
-
-    const containerStyle: React.CSSProperties = {
-      position: 'relative',
-      background: 'white',
-      padding: '0.5rem',
-      borderRadius: '8px',
-      boxShadow: '0 1px 3px rgba(0,0,0,0.1)',
-      width: '708px',
-      minHeight: '400px',
-      boxSizing: 'border-box',
-      flexShrink: 0,
-      display: 'flex',
-      flexDirection: 'column',
-      border: tableColor ? `2px solid ${tableColor}20` : undefined
-    }
-
-    const countIndicator = countIndicatorOverride ?? renderTableCountIndicator(tableName, field, cacheKey)
-
-    const metricLabels = getMetricLabels(aggregation)
-    const pathLabel = formatMetricPath(aggregation)
-
-    // Map state values to codes for Plotly choropleth
-    // Aggregate counts by state code (to handle both "CA" and "California")
-    const stateMap = new Map<string, { count: number, name: string, originalValues: any[] }>()
-
-    aggregation.categories.forEach(category => {
-      const stateValue = category.value === '' ? '(Empty)' : String(category.value)
-      const normalizedName = normalizeStateName(stateValue)
-      const stateCode = getStateCode(normalizedName)
-
-      if (stateCode) {
-        const existing = stateMap.get(stateCode)
-        if (existing) {
-          existing.count += category.count
-          existing.originalValues.push(normalizeFilterValue(category.value))
-        } else {
-          stateMap.set(stateCode, {
-            count: category.count,
-            name: normalizedName,
-            originalValues: [normalizeFilterValue(category.value)]
-          })
-        }
-      }
-    })
-
-    const locationCodes: string[] = []
-    const zValues: number[] = []
-    const hoverTexts: string[] = []
-    const filterValues: any[][] = []
-
-    stateMap.forEach((data, code) => {
-      locationCodes.push(code)
-      zValues.push(data.count)
-      hoverTexts.push(data.name)
-      filterValues.push(data.originalValues)
-    })
-
-    if (locationCodes.length === 0) {
-      // No valid US state data - fall back to categorical table view
-      return renderTableView(title, tableName, field, tableColor, aggregationOverride, cacheKeyOverride, countIndicatorOverride)
-    }
-
-    const totalCount = aggregation.total_rows ?? zValues.reduce((sum, val) => sum + val, 0)
-
-    const baselineAggregation = getBaselineAggregation(tableName, field)
-    const categoriesForMenu =
-      metricsMatch(baselineAggregation, aggregation) && baselineAggregation?.categories?.length
-        ? baselineAggregation.categories
-        : aggregation.categories
-
-    return (
-      <div style={containerStyle}>
-        {renderChartHeader({
-          title: metadata?.display_name || title,
-          tooltip: tooltipText,
-          countIndicator,
-          actions: actionButtons,
-          isListColumn: metadata?.is_list_column || false
-        })}
-        <Plot
-          data={[{
-            type: 'choropleth',
-            locationmode: 'USA-states',
-            locations: locationCodes,
-            z: zValues,
-            text: hoverTexts,
-            hovertemplate: `${['%{text}', `Count (${metricLabels.short}): %{z}`, 'Percent: %{customdata}%']
-              .concat(pathLabel ? [pathLabel] : [])
-              .join('<br>')}<extra></extra>`,
-            customdata: zValues.map(val =>
-              totalCount > 0 ? ((val / totalCount) * 100).toFixed(1) : '0'
-            ),
-            colorscale: [
-              [0, tableColor ? `${tableColor}40` : '#E3F2FD'],
-              [1, tableColor || '#2196F3']
-            ],
-            marker: {
-              line: {
-                color: locationCodes.map((_code, idx) =>
-                  filterValues[idx].some(v => isValueFiltered(field, v, cacheKey)) ? '#000' : 'white'
-                ),
-                width: locationCodes.map((_code, idx) =>
-                  filterValues[idx].some(v => isValueFiltered(field, v, cacheKey)) ? 3 : 1
-                )
-              }
-            },
-            showscale: true,
-            colorbar: {
-              title: metricLabels.short,
-              titleside: 'right',
-              tickfont: { size: 10 },
-              len: 0.7
-            }
-          }]}
-          layout={{
-            geo: {
-              scope: 'usa',
-              projection: { type: 'albers usa' },
-              showlakes: true,
-              lakecolor: 'rgb(255, 255, 255)'
-            },
-            height: 380,
-            margin: { t: 5, b: 5, l: 5, r: 5 },
-            paper_bgcolor: 'transparent',
-            dragmode: false
-          }}
-          config={{
-            displayModeBar: false,
-            responsive: true,
-            staticPlot: false,
-            scrollZoom: false
-          }}
-          style={{ width: '698px', height: '380px', cursor: 'pointer' }}
-          onClick={(event: PlotMouseEvent) => {
-            const point = event.points?.[0]
-            if (!point) return
-
-            const pointIndex = point.pointIndex
-            if (typeof pointIndex === 'number' && pointIndex >= 0 && pointIndex < filterValues.length) {
-              const stateValues = filterValues[pointIndex]
-              // If state has multiple representations (e.g., "CA" and "California"), handle as multi-value
-              if (stateValues.length === 1) {
-                toggleFilter(field, stateValues[0], tableName, cacheKey)
-              } else {
-                // Create/toggle filter with all state representations
-                setFilters(prev => [
-                  ...removeColumnFilters(prev, field, cacheKey),
-                  { column: field, operator: 'in', value: stateValues, tableName, countByKey: cacheKey }
-                ])
-              }
-            }
-          }}
-          onSelected={(event: PlotSelectionEvent) => {
-            if (!event?.points || event.points.length === 0) return
-            const selectedValues = event.points
-              .map(p => p.pointIndex)
-              .filter((idx): idx is number => typeof idx === 'number' && idx >= 0 && idx < filterValues.length)
-              .flatMap(idx => filterValues[idx]) // Flatten all state value arrays
-
-            if (selectedValues.length > 0) {
-              setFilters(prev => [
-                ...removeColumnFilters(prev, field, cacheKey),
-                { column: field, operator: 'in', value: selectedValues, tableName, countByKey: cacheKey }
-              ])
-            }
-          }}
-        />
-        {renderFilterMenu(tableName, field, categoriesForMenu, cacheKey)}
-      </div>
-    )
+  const chartContextValue = {
+    getEffectiveCacheKeyForChart,
+    getAggregation,
+    getBaselineAggregation,
+    getMetricLabels,
+    metricsMatch,
+    getColumnMetadata,
+    getTableDisplayNameByName,
+    formatMetricPath,
+    getCountByOptions,
+    getCountByCacheKey,
+    handleCountByChange,
+    handleChartCountOverrideChange,
+    activeCountMenuKey,
+    setActiveCountMenuKey,
+    getTableColor,
+    getCountIndicatorColor,
+    getCountByTableColor,
+    getCountByLabelFromCacheKey,
+    activeFilterMenu,
+    setActiveFilterMenu,
+    filters,
+    setFilters,
+    hasColumnFilter,
+    isValueFiltered,
+    normalizeFilterValue,
+    toggleFilter,
+    clearColumnFilter,
+    removeColumnFilters,
+    rangeSelections,
+    customRangeInputs,
+    toggleRangeFilter,
+    isRangeFiltered,
+    handleCustomRangeChange,
+    applyCustomRange,
+    updateColumnRanges,
+    formatRangeValue,
+    getDisplayHistogram,
+    showPercentageLabels,
+    toggleViewPreference,
+    getSurvivalViewPreference,
+    toggleSurvivalViewPreference,
+    findTable: (tableName: string) => dataset?.tables.find(t => t.name === tableName),
+    findSurvivalStatusColumn,
+    ensureSurvivalCurve,
+    getSurvivalCurve,
+    isOnDashboard,
+    toggleDashboard,
   }
 
   if (loading) return <p>Loading explorer...</p>
@@ -3677,6 +1763,7 @@ function DatasetExplorer() {
   if (!dataset) return <p>Dataset not found</p>
 
   return (
+    <ChartProvider value={chartContextValue}>
     <div>
       {/* Header */}
       <div style={{ marginBottom: '2rem', background: 'white', padding: '1.5rem', borderRadius: '8px', boxShadow: '0 1px 3px rgba(0,0,0,0.1)', position: 'relative' }}>
@@ -5023,21 +3110,8 @@ function DatasetExplorer() {
 
                     if (viewPref === 'table') {
                       return (
-                        <div
-                          key={cardKey}
-                          ref={cardRef}
-                          data-dashboard-key={cardKey}
-                          style={{ gridColumn: 'span 2', gridRow: 'span 2' }}
-                        >
-                          {renderTableView(
-                            displayTitle,
-                            tableName,
-                            columnName,
-                            tableColor,
-                            aggregation,
-                            overrideKey,
-                            indicatorNode
-                          )}
+                        <div key={cardKey} ref={cardRef} data-dashboard-key={cardKey} style={{ gridColumn: 'span 2', gridRow: 'span 2' }}>
+                          <TableViewChart title={displayTitle} tableName={tableName} field={columnName} tableColor={tableColor} aggregationOverride={aggregation} cacheKeyOverride={overrideKey} countIndicatorOverride={indicatorNode} />
                         </div>
                       )
                     }
@@ -5045,30 +3119,14 @@ function DatasetExplorer() {
                     if (allowPie) {
                       return (
                         <div key={cardKey} ref={cardRef} data-dashboard-key={cardKey}>
-                          {renderPieChart(
-                            displayTitle,
-                            tableName,
-                            columnName,
-                            tableColor,
-                            aggregation,
-                            overrideKey,
-                            indicatorNode
-                          )}
+                          <PieChart title={displayTitle} tableName={tableName} field={columnName} tableColor={tableColor} aggregationOverride={aggregation} cacheKeyOverride={overrideKey} countIndicatorOverride={indicatorNode} />
                         </div>
                       )
                     }
 
                     return (
                       <div key={cardKey} ref={cardRef} data-dashboard-key={cardKey} style={{ gridColumn: 'span 2' }}>
-                        {renderBarChart(
-                          displayTitle,
-                          tableName,
-                          columnName,
-                          tableColor,
-                          aggregation,
-                          overrideKey,
-                          indicatorNode
-                        )}
+                        <BarChart title={displayTitle} tableName={tableName} field={columnName} tableColor={tableColor} aggregationOverride={aggregation} cacheKeyOverride={overrideKey} countIndicatorOverride={indicatorNode} />
                       </div>
                     )
                   } else if (metaDisplayType === 'survival_time') {
@@ -5104,61 +3162,26 @@ function DatasetExplorer() {
                     if (view === 'km') {
                       return (
                         <div key={cardKey} ref={cardRef} data-dashboard-key={cardKey} style={{ gridColumn: 'span 2', gridRow: 'span 2' }}>
-                          {renderSurvivalChart(
-                            displayTitle,
-                            tableName,
-                            columnName,
-                            tableColor,
-                            aggregation,
-                            overrideKey,
-                            indicatorNode,
-                            toggleButton,
-                            false
-                          )}
+                          <SurvivalChart title={displayTitle} tableName={tableName} field={columnName} tableColor={tableColor} aggregationOverride={aggregation} cacheKeyOverride={overrideKey} countIndicatorOverride={indicatorNode} extraActions={toggleButton} showHistogram={false} />
                         </div>
                       )
                     }
 
                     return (
                       <div key={cardKey} ref={cardRef} data-dashboard-key={cardKey} style={{ gridColumn: 'span 2' }}>
-                        {renderHistogram(
-                          displayTitle,
-                          tableName,
-                          columnName,
-                          tableColor,
-                          aggregation,
-                          overrideKey,
-                          indicatorNode,
-                          toggleButton
-                        )}
+                        <HistogramChart title={displayTitle} tableName={tableName} field={columnName} tableColor={tableColor} aggregationOverride={aggregation} cacheKeyOverride={overrideKey} countIndicatorOverride={indicatorNode} extraActions={toggleButton} />
                       </div>
                     )
                   } else if (normalizedDisplayType === 'numeric' && aggregation.histogram) {
                     return (
                       <div key={cardKey} ref={cardRef} data-dashboard-key={cardKey} style={{ gridColumn: 'span 2' }}>
-                        {renderHistogram(
-                          displayTitle,
-                          tableName,
-                          columnName,
-                          tableColor,
-                          aggregation,
-                          overrideKey,
-                          indicatorNode
-                        )}
+                        <HistogramChart title={displayTitle} tableName={tableName} field={columnName} tableColor={tableColor} aggregationOverride={aggregation} cacheKeyOverride={overrideKey} countIndicatorOverride={indicatorNode} />
                       </div>
                     )
                   } else if (aggregation.display_type === 'geographic' && aggregation.categories) {
                     return (
                       <div key={cardKey} ref={cardRef} data-dashboard-key={cardKey} style={{ gridColumn: 'span 4' }}>
-                        {renderMapChart(
-                          displayTitle,
-                          tableName,
-                          columnName,
-                          tableColor,
-                          aggregation,
-                          overrideKey,
-                          indicatorNode
-                        )}
+                        <MapChart title={displayTitle} tableName={tableName} field={columnName} tableColor={tableColor} aggregationOverride={aggregation} cacheKeyOverride={overrideKey} countIndicatorOverride={indicatorNode} />
                       </div>
                     )
                   }
@@ -5402,7 +3425,7 @@ function DatasetExplorer() {
                     if (viewPref === 'table') {
                       return (
                         <div key={`${table.name}_${agg.column_name}`} style={{ gridColumn: 'span 2', gridRow: 'span 2' }}>
-                          {renderTableView(displayTitle, table.name, agg.column_name, tableColor, aggregationForChart, cacheKey)}
+                          <TableViewChart title={displayTitle} tableName={table.name} field={agg.column_name} tableColor={tableColor} aggregationOverride={aggregationForChart} cacheKeyOverride={cacheKey} />
                         </div>
                       )
                     }
@@ -5410,14 +3433,14 @@ function DatasetExplorer() {
                     if (allowPie) {
                       return (
                         <div key={`${table.name}_${agg.column_name}`}>
-                          {renderPieChart(displayTitle, table.name, agg.column_name, tableColor, aggregationForChart, cacheKey)}
+                          <PieChart title={displayTitle} tableName={table.name} field={agg.column_name} tableColor={tableColor} aggregationOverride={aggregationForChart} cacheKeyOverride={cacheKey} />
                         </div>
                       )
                     }
 
                     return (
                       <div key={`${table.name}_${agg.column_name}`} style={{ gridColumn: 'span 2' }}>
-                        {renderBarChart(displayTitle, table.name, agg.column_name, tableColor, aggregationForChart, cacheKey)}
+                        <BarChart title={displayTitle} tableName={table.name} field={agg.column_name} tableColor={tableColor} aggregationOverride={aggregationForChart} cacheKeyOverride={cacheKey} />
                       </div>
                     )
                   } else if (metaDisplayType === 'survival_time') {
@@ -5453,26 +3476,26 @@ function DatasetExplorer() {
                     if (view === 'km') {
                       return (
                         <div key={`${table.name}_${agg.column_name}_km`} style={{ gridColumn: 'span 2', gridRow: 'span 2' }}>
-                          {renderSurvivalChart(displayTitle, table.name, agg.column_name, tableColor, aggregationForChart, cacheKey, undefined, toggleButton, false)}
+                          <SurvivalChart title={displayTitle} tableName={table.name} field={agg.column_name} tableColor={tableColor} aggregationOverride={aggregationForChart} cacheKeyOverride={cacheKey} extraActions={toggleButton} showHistogram={false} />
                         </div>
                       )
                     }
 
                     return (
                       <div key={`${table.name}_${agg.column_name}_hist`} style={{ gridColumn: 'span 2' }}>
-                        {renderHistogram(displayTitle, table.name, agg.column_name, tableColor, aggregationForChart, cacheKey, undefined, toggleButton)}
+                        <HistogramChart title={displayTitle} tableName={table.name} field={agg.column_name} tableColor={tableColor} aggregationOverride={aggregationForChart} cacheKeyOverride={cacheKey} extraActions={toggleButton} />
                       </div>
                     )
                   } else if (normalizedDisplayType === 'numeric' && agg.histogram) {
                     return (
                       <div key={`${table.name}_${agg.column_name}`} style={{ gridColumn: 'span 2' }}>
-                        {renderHistogram(displayTitle, table.name, agg.column_name, tableColor, aggregationForChart, cacheKey)}
+                        <HistogramChart title={displayTitle} tableName={table.name} field={agg.column_name} tableColor={tableColor} aggregationOverride={aggregationForChart} cacheKeyOverride={cacheKey} />
                       </div>
                     )
                   } else if (agg.display_type === 'geographic' && agg.categories) {
                     return (
                       <div key={`${table.name}_${agg.column_name}`} style={{ gridColumn: 'span 4' }}>
-                        {renderMapChart(displayTitle, table.name, agg.column_name, tableColor, aggregationForChart, cacheKey)}
+                        <MapChart title={displayTitle} tableName={table.name} field={agg.column_name} tableColor={tableColor} aggregationOverride={aggregationForChart} cacheKeyOverride={cacheKey} />
                       </div>
                     )
                   }
@@ -5483,6 +3506,7 @@ function DatasetExplorer() {
           )
         })}
     </div>
+    </ChartProvider>
   )
 }
